Set description, deprecation_reason on mutations. (#1697)
This sets the `description` of a GraphQL mutation to its class' docstring, which makes it easier to see the documentation in GraphiQL.

It also sets the `deprecation_reason` to its class' attribute of the same name, if it's defined on it.  This allows our mutations to be deprecated.

GraphQL for deprecated mutations isn't generated by the querybuilder's auto-generator.

This also adds a `yarn graphql` script which regenerates `schema.json` and then runs the querybuilder.