Ignore empty GraphQL files. (#2010)
When manually creating new GraphQL files in the `frontend/lib/queries` directory, the process can be annoying because one first (usually) creates an empty `.graphql` file in that directory, then populates it and saves it.  However, creating the empty file triggers a distracting syntax error in querybuilder's watcher, because it tries to open the file and parse it as GraphQL and finds nothing.

This ignores empty `.graphql` files so such errors don't occur anymore.