Add an admin user directory (#2018)
This is an attempt at adding a user directory with super-friendly autocomplete to the front-end, which would be useful to embed in the Front app plugin.

(This doesn't actually do the embedding in the Front app plugin, that will be done in a future PR.)

It currently adds a very simple user directory at `/admin/directory/` with an autocomplete.  When choosing an autocomplete suggestion, the same "user information" widget that shows in the admin conversations view and Front app plugin is displayed.

From a technical standpoint, the main challenge here was actually accessing a GraphQL endpoint for the search results: until now, all our autocomplete widgets have used a third-party REST API to fetch suggestions.  I actually ended up just forking the `SearchRequester` implementation from `@justfixnyc/geosearch-requester`, calling the new class `GraphQLSearchRequester`.  I don't feel great about this and filed https://github.com/JustFixNYC/justfix-ts/issues/25 to address it later.