Add a user menu to navbar. (#2026)
This adds a user menu to the app.justfix.nyc navbar that consists of the user's initials with a circle around them (if the user lacks a first/last name, a ⚙ is shown instead).  The "log out" menu item has been moved under this menu, and an "account settings" link has been added to it.   For staff users, the "admin" link has been moved under this menu too.

The NoRent and EvictionFreeNY navbars remain unchanged, though it's likely that we'll eventually show this menu on them too.