Show city/state in national address. (#2003)
This is another attempt to address the UX issues mentioned in #2002.  It adds a read-only "city and state" field to the national address step that reminds the user of what they entered in the previous step, and how it relates to what they're entering now.  A "change" link appears to the city and state; clicking it takes the user back to the previous step.