Make email optional for LOC. (#2041)
This makes the user's email optional for the LOC flow.  Because it was originally designed without even asking for email, doing this was pretty straightforward--we just had to make it possible for email to be optional in the final step of onboarding for LOC (but not for EHP, since it's required for e-signing there).

Also, if the user didn't enter an email address, the "Thanks for signing up, an email to verify your account is on its way" screen is never shown.

This also introduces new helper functions to make it easier to add the localized text ` (optional)` to labels, and refactors some existing code to use it.

Finally, it also gets rid of a few deprecated forms that weren't being used anymore.