Remove "en-us" locale key from efnycURL function (#2048)
Now that we no longer direct users to the old EFNYC site, we no longer need to force our locale key to be en-us when linking to the English version of the site. Now, we direct folks to Eviction Free NY, which is part of the tenant platform and uses the same language keys.