Add user admin tabs for action-specific details (#2049)
This adds tabs to the top of the user admin pages that link to the action-specific admin pages for each user, and show which one is currently being viewed, along with basic information on what stage of progress an action is in.

The levels of progress are "not started", "in progress", and "completed".  If an action hasn't yet been started, its tab will be gray.