Fix MultipleObjectsReturned on user change admin. (#2053)
Urf, it looks like #2049 caused a regression whereby visiting user admin pages for users that had sent multiple NoRent letters triggered a `MultipleObjectsReturned` error.  It looks like this is due to the fact that our annotations that involves one-to-many relationships from the user model were actually causing multiple results per user to be returned (e.g., a separate user record for each NoRent letter).

This fixes the problem, along with an analogous one for HP actions.