Fix server-side development bundles. (#2068)
I noticed on local development that when visiting `/admin/directory/`, I got a server-side exception that was being raised because `cuid` (a dependency) was expecting a browser environment.

Interestingly, this error wasn't occurring in production or staging.

I noticed that the instructions for using webpack 5 with `webpack-node-externals` specified to _remove_ `target: 'node'` from our node webpack configuration, so I did that when upgrading to webpack 5 in #2028, and I suspected that might have something to do with `cuid` expecting a browser environment.  Sure enough, when I added `target: 'node'` back to the configuration, everything worked fine again.  I've since filed this as https://github.com/liady/webpack-node-externals/issues/104.

However, upon further contemplation, I realized that _all our node dependencies_ were being included in our `lambda.js` bundle--i.e., the `webpack-node-externals` package wasn't actually doing anything helpful!  I'm not sure if this was happening with webpack 4 as well, but the root cause was because all our dependencies in local development are actually in `/node_modules` (rather than the usual `./node_modules` folder relative to the location of `package.json`).  The reason that the exception wasn't being raised on production or staging was because in those environments, all our dependencies were actually where webpack expected them to be (and since it was node loading the dependencies, it never loaded any versions that expected a browser environment).  Once I told `webpack-node-externals` to look in `/node_modules` in our development environment, everything worked as expected.

So, this PR fixes our node webpack configuration by re-introducing `target: 'node'`, and it also ensures that none of our third-party dependencies (except for `@justfixnyc` packages, which need to be transpiled) are included in the server-side bundles during development.