Replace May deadline for hardship declarations with projected August one on EFNY (#2062)
This PR updates any mention of the state-wide deadline for submitting a hardship declaration form to August 31st. After consulting with Right to Counsel and others, it became clear that we should also include a note on the landing page to notify folks of the new extended moratorium, as well as a disclaimer within the builder flow to explain that while the old form still says "May 1" on it, the protections from the form last until August.