Add a new ClearAnonymousSession mutation and use it. (#2081)
In EFNY, NoRent, and Rent History online's welcome page, as well as on step 1 of onboarding, if the user clicks the "cancel" button and are logged in, they are actually logged out, which is unexpected and disorienting.  This is happening because the button is actually executing the `Logout` GraphQL mutation!

What we really need is for the "cancel" button to reset the user's session _if they're not logged in_ (i.e., if the session is anonymous) and to do nothing otherwise (other than navigate the user's browser to a different page).

So, this PR adds a new `ClearAnonymousSession` mutation that does just that.