Add support for multiple versions of the EFNY PDF. (#2075)
It looks like there's going to be a new version of the EFNY Hardship Declaration PDF that (hopefully) just changes the date at which protections end.

We can't just simply replace the PDF on our end, because we still need to be able to render old PDFs (ones that were e.g. signed in February) in their original form.  So, this adds a new `pdf_version` attribute to our `HardshipDeclarationVariables` object (which is serialized to the database as JSON) that records what version of the declaration the user signed.

We ensure backwards compatibility by defaulting its value to `1` (the original declaration PDF) when it's not present--this is so that when we pull out legacy `HardshipDeclarationVariables` instances that we are pulling from `SubmittedHardshipDeclaration` models, they are deserialized in a way that represents their true state.

Note that since this only changes the schema of `HardshipDeclarationVariables`--an object that is stored as JSON, not as a database row--this change does not require a database migration (i.e., we're using schema-on-read semantics here).

This also removes the "disclaimer" about the "May 1st" deadline on the agree to terms page, and updates a few outbound links (these contributions were made by @sraby).