Add django-sql-dashboard integration. (#2103)
Throughout the lifetime of the Tenant Platform, we've frequently had the need to be able to run arbitrary SQL queries against our database to gain insight into how our software is used:

* The vega-powered dashboard added in #568 was an attempt to make gaining insights more accessible, but suffers from not being very flexible.  For example, adding a new visualization requires committing a change to the codebase and deploying it, which is inconvenient.  Even then, the visualization isn't necessarily easy to explore or filter--for example, zooming-in to a particular time range is impossible.

* We've generated one-off SQL queries for individual tasks and recorded their SQL for posterity in issues like #1998.  However, they require a developer with full access to our production database to run.

* We export _some_ of our backend data to Amplitude (see #1948), where it can be thoroughly analyzed.  However, we have to be very careful about how we export data to that platform, as it's append-only, and even Amplitude's UI only allows certain kinds of queries.

* SQL is learnable by non-engineers, and it's unfortunate that more of our staff don't have the ability to harness it to gain insights about how our products are used.

This alleviates the above concerns by fixing #2011 and giving us a handy and secure SQL dashboard via the integration of [django-sql-dashboard][].

## Configuration

As per the "everything is an optional feature" philosophy of this codebase, the dashboard is disabled by default--that is, the admin will not link to it, and attempts to visit it by URL will 404. To enable it, one needs to set the `DASHBOARD_DATABASE_URL` environment variable.  Local docker-compose based dev setups can set this to `postgis://justfix@db/justfix`, the URL of the local dev database, to quickly try it out--though this will give SQL access to the entire database, which is not a good idea for production (for that we'll likely want to set up a [read-only Postgres user][] with access only to allowed tables; regardless, this PR is independent of such details, as they're all dictated by the environment).

In order for non-superusers to access the dashboard, they will need to be staff members with the `django_sql_dashboard.execute_sql` permission.  For now I've given all outreach coordinators this permission, since we plan on making available only the tables/columns containing data that they already have access to via other means (e.g. admin views and data downloads).

## Implementation notes

Implementation wasn't super straightforward since the dashboard doesn't want to be mounted off the `admin` namespace, which makes it a bit harder to decorate it with standard admin protections.  However, I did manage to accomplish it, along with unit tests that verify the dashboard endpoints are protected the way all our other admin endpoints are (e.g. requiring 2FA to access).

## Relation to data downloads

Finally, it should be noted that we still have a "data downloads" feature (see #521 for a screenshot) that is available to outreach coordinators and partners (see #1764, #1929, #1913 for examples).  I think of those as orthogonal from this dashboard, because they still provide important functionality that isn't fully addressed by django-sql-dashboard, such as:

* Built-in data dictionaries (thanks to #1935)
* Strong, finely-grained access permission control (partners have access _only_ to download the specific dataset we've given them access to; I'm wary of giving them blanket `django_sql_dashboard.execute_sql` permission)
* Source-controlled version history, so we can easily see how the data we make available to partners has evolved over time

[django-sql-dashboard]: https://github.com/simonw/django-sql-dashboard
[read-only Postgres user]: https://django-sql-dashboard.datasette.io/en/latest/setup.html#setting-up-read-only-postgresql-credentials

## Deprecation of the old admin dashboard

This PR formally deprecates the aforementioned vega-powered dashboard in favor of django-sql-dashboard.  The old dashboard is still accessible at `/admin/vega-dashboard/` but it isn't linked to from anywhere, and once we merge this PR I will create a new issue to remove it from the repository.  (I didn't fully remove it in this PR partly because I wanted to keep this PR as simple as possible, and partly because there may actually be some folks who use it that I'm not aware of.)