Add a twilio.SendSmsResult class. (#2108)
This tees us up for fixing #2057 and addressing edge cases identified in #2076 by making `twilio.send_sms()` and `JustfixUser.send_sms()` return a new `SendSmsResult` dataclass instead of a string.

This is because simply returning an empty string in the case of failing to send and SMS isn't enough: for us to figure out what to do next: in the case of a network error, for example, we'll want to retry the sending at a later date, but in the case of an invalid phone number, we won't.  This new return type gives clients enough information to make that decision.