Remember non-retryable errors for SMS reminders. (#2111)
This helps with #2076 by altering the `texting.Reminder` model to remember when we've attempted to send a SMS reminder that failed in a non-retryable way, e.g. because the user blocked us or somesuch.  This will ensure that we don't needlessly try to re-send the reminder in the future.

It also adds a `--seconds-between-texts` option to the `send_norent_reminders` command, which will help with #2057 by making it possible for us to manually throttle the rate at which we give SMS messages to Twilio, reducing the possibility that we will run into "SMS queue is full" errors.