Disable entry to EHP tool while housing court is denying cases (#2118)
Given that the courts have blocked our EHP tool from helping tenants sue their landlord, this quick fix blocks entry points to the tool from DDO, the splash page, and the welcome page. We also link folks to a google form that helps them get a legal referral.

The rest of the team was made aware that this will NOT block users within the EHP flow from completing the process, nor will it change the confirmation page for folks who have already finished using the tool. Program team support and others are stepping in to fill these gaps.