Add NYCx app w/ single endpoint. (#2128)
This borrows liberally from Sam's https://github.com/JustFixNYC/who-owns-what/pull/469 to provide a NYCx Django app and `/nycx/address` endpoint for use by the NYCx prototype textbot. 

Note that the aforementioned PR was closed without merging; the explanation is at https://github.com/JustFixNYC/who-owns-what/pull/469#issuecomment-861586136.

However, things have changed since then!  We decided to use TextIt as our platform for prototyping instead of DialogFlow--see [Clubhouse story #7453](https://app.clubhouse.io/justfixnyc/story/7453/consider-using-textit-instead-of-dialogflow-for-the-repairs-flow) for details on why--and this meant that we had no existing fullfillment server in which to put an SQL query.  Or indeed, none from which we could do geocoding, either.

This resolves the prototype's lack of a server by adding a new Django app to the tenant platform called `nycx`.  This app currently only provides a single API endpoint, which does _both_ geocoding _and_ housing type prediction in one fell swoop.

The API is documented at `/nycx/`.

## Why is this on the tenant platform?

You'll note that the aforementioned PR was on WoW, not the tenant platform.  I decided to put it on the TP instead for a number of reasons:

* The TP has a writable database, which means that we have a place for persistent storage if we need it.  Also, Django's app infrastructure is such that NYCx's data will be kept compartmentalized from the rest of the TP's data--once we're done prototyping, if we don't need something, it will be easy to delete all the NYCx data without accidentally deleting other TP data (for instance, NYCx-only data can easily be wiped via `python manage.py migrate nycx zero`).

* The TP has direct access to user data. We don't yet know if we want NYCx to have access to user accounts--for instance, we might eventually want to make it possible to send a LOC from the textbot, or perhaps we might want to detect a user's housing type from the one they specified during onboarding--so housing the NYCx endpoints in the TP will give us easy access to that.

* The TP already has a bunch of server-side functionality built-in, such as `project.geocoding`.  At the moment, our coding infrastructure is set up in such a way that cross-repository reuse of Python code isn't easy, so it's more advantageous to put Python code that needs to leverage our existing Python code in the TP rather than WoW.