Perform sanity check after creating devcontainer. (#2133)
@samaratrilling was running into problems where VSCode would not associate its python interpreter with `/venv/bin/python`, and instead use the container's system-wide installation of Python, which wouldn't have any third-party libraries installed.

This was likely happening because at some point, `docker-compose down -v` was run, which destroyed `/venv`.  Once VSCode saw that the Python interpreter it was supposed to use didn't exist, it likely silently switched over to the system-wide one.

This adds a "sanity check" after creating the container which verifies that `/venv/bin/python` exists.  A terminal window is opened after the container is built that tells us if everything is copacetic.

Note that this only happens when the container is _created_, which means that if `docker-compose down -v` is subsequently run, VSCode will silently fall back to using a different interpreter.  The only way to check more often is to make the sanity check a [`postStartCommand`](https://code.visualstudio.com/docs/remote/devcontainerjson-reference#_lifecycle-scripts), which would run the sanity check every time VSCode is started--however, because this _always_ opens the terminal window, I thought it would get annoying, so I limited it to running on container creation.