Redirect if a path can be resolved by adding or removing a slash (#2136)
I've seen myself and others manually type a pathname into the site and have it return a 404 because of a missing or extra `/`.

This is exacerbated by the fact that our routes don't actually have a very consistent convention.  For example, to go to the latest step in a Letter of Complaint, the path is `/loc` (without a trailing slash), whereas to go to the development tools page, it's `/dev/` (with a trailing slash).

This modifies our site routing logic such that, before returning a 404 page, we first check to see if adding or removing a trailing slash results in a valid path.  If it does, we redirect the user to it instead of 404'ing.

Note that the act of appending a slash is actually consistent with Django's default behavior, since its [`APPEND_SLASH`][] setting defaults to `True`; Django just can't "see" our React-based routes, so it has no way of knowing whether to append a slash or not.

On the other hand _removing_ a trailing slash is not something Django normally does, but I don't see any drawbacks to doing so.

[`APPEND_SLASH`]: https://docs.djangoproject.com/en/dev/ref/settings/?from=olddocs#append-slash