Rename NorentScaffolding to OnboardingScaffolding in backend. (#2150)
This attempts to renamed `NorentScaffolding` to `OnboardingScaffolding` _only in the backend implementation_, without actually making any changes to GraphQL APIs.  At the very least, it decouples `onboarding` from `norent`.

It also centralizes a bunch of scaffolding-related logic in `onboarding.scaffolding`, rather than solely the Pydantic model.  That is, the module now includes the GraphQL type for scaffolding, as well as a number of utility classes to make GraphQL stuff easier.

This will help inch our way towards #2142.  It's also a small subset of the abandoned #2145.