Pull content for moratorium banner from Contentful. (#2125)
This is an attempt to try pulling content for the Covid moratorium banner text directly from Contentful, so that we don't need to push a change to this repository whenever we want to change the text of the banner.

It currently uses the new [`@justfixnyc/contentful-common-strings`](https://www.npmjs.com/package/@justfixnyc/contentful-common-strings) package to help accomplish this.

## How it works

The tenant platform now optionally supports retrieving "common strings" from Contentful.  By "common strings" we really mean "common [rich text](https://www.contentful.com/developers/docs/concepts/rich-text/) strings".

Each common string has an alphanumeric id and a value.  The value is a localizable rich text value.  For example, in this PR, the common string with id `covidMoratoriumBanner` is expected to have a value containing the rich text to show users about the current state of the COVID Moratorium, localized in both English and Spanish.

The platform retrieves all entries from its configured Contentful space [tagged](https://www.contentful.com/help/tags/) `common`.  Each of these entries is expected to have a short text field called `id` and a localized rich text field called `value`.  

The common strings are made available to front-end code,.  A helper function on the front-end makes it convenient to retrieve a common string's rich text document representation by its id.  The front-end can then render them using the `@contentful/rich-text-react-renderer` package.

Note that at present, _all_ locales of the common strings are passed to the front-end, which means that there's a slight amount of bloat--but it shouldn't be _too_ bad since we have very few common strings.

The back-end caches the common strings from Contentful for a short period of time--at the time of this writing, it's 5 seconds--to ensure minimal latency during periods of high load.  (Note that the back-end is responsible for retrieving the Contentful strings to ensure that they are properly rendered during server-side rendering.)