Track the front-end version in Amplitude. (#2157)
We don't really have a way to track what users are on what versions of our front-ends right now.  I think that Rollbar technically supports this kind of thing, but as far as I can tell, it is not supported on their free plan (which is what we use).  So, this PR adds a new `frontendVersion` property to our Amplitude user properties, which is the git revision used to build the front-end.

This will hopefully make it easier to figure out when it's safe to remove deprecated GraphQL endpoints and such.