In backend, use onboarding scaffolding to store steps 1 and 3. (#2154)
This further moves us towards #2142 by internally storing steps 1 and 3 of app.justfix.nyc onboarding data in our database.  **It still does not change anything in the front-end.**  It just makes submitting onboarding steps change the scaffolding, and retrieving data for onboarding steps 1 and 3 from the front-end actually returns data from scaffolding in the backend.  Legacy data stored in the old location is also migrated to the scaffolding as needed.

In a future PR, we will deprecate `session.onboardingStep1` and `session.onboardingStep3` and instead only look at the scaffolding data.

We also account for the weird use of onboarding step 1 for only address information that was deprecated in #2143 and migrate the data accordingly.