Use PostGIS/GeoDjango to see if lnglat is in NYC. (#2166)
Fixes #2152.  Our test suite should actually take a tiny bit shorter to run now, too.

It should be noted that in order for this to work on local dev setups (outside of just running the test suite), you'll need to run `manage.py loadfindhelpdata`.  This is mentioned in the README but I wanted to call it out here.  Also, it's automatically run every time we push a deployment (the command is idempotent and will update any data that needs updating).