Add session.onboardingScaffolding to GraphQL schema. (#2160)
This moves us closer to #2142 by adding `session.onboardingScaffolding` to the GraphQL schema and deprecating the following fields in favor of it:

* `session.norentScaffolding`
* `session.onboardingStep1`
* `session.onboardingStep3`
* `session.rentalHistoryInfo`

Note that some back-end tests are still referring to some of these deprecated fields.  I didn't change them in part because this PR was becoming a bit larger than I realized and I figured we could also take care of them when we actually remove the deprecated fields (at the very latest).