Disallow EFNY declarations to be sent on the server-side. (#2178)
This is a complement to #2177 which disallows users to send EFNY declarations on the server-side. Attempting to submit their EFNY declaration will give them a form error with the text:

This tool has been discontinued! Please reload the page for more details.

This will ensure that users on old versions of the front-end still won't be able to send declarations.

Note that telling the user to reload their page assumes that #2177 has been merged. Since reloading the user's browser will then trigger a redirect to the EFNY homepage with a message telling the user why they can't proceed.