Make shortlink slug always default to lowercase (#2185)
This PR updates our short links generator on the Tenant Platform such that:

All shortlink slugs are saved as lowercase (i.e if someone tries to create a short link called app.justfix.nyc/s/HCA it will be saved as .../hca)
If a user includes a mix of cases when typing a short URL, they will be redirected to an all-lowercase version of that URL (I.e. if someone then tries to access app.justfix.nyc/s/HCA they will be redirected to .../hca)
My sense is that since this is only affecting the save() function on the Link model, this will not actually require a database migration and keep all existing links intact. Luckily, though, all of the links created in production are fully lowercase to begin with!