Add feature flag to suspend EFNY functionality (#2192)
This PR creates a new env variable called IS_EFNY_SUSPENDED that defaults to false but when set to true will suspend all functionality on EvictionFree that allows user to create and send out hardship declaration. It borrows code from old commits that originally suspended the EvictionFree tool a few weeks ago (specifically, 0d2d2ef, and and31e878708ced25b1454f0f1312138a1ae31e2822). Now, we will be able to revert back to this suspended state of the tool any time by just setting this new env variable.

NOTE: to avoid complexity, I decided to leave the about, faq, and confirmation pages unchanged even when Eviction Free NY is suspended. In the past, we removed the first two pages entirely, and changed a lot of the text of the confirmation page. On this second look, I thought that these changes were unnecessary so I kept things as is.