Fix CSS issue with Build My Declaration button (#2197)
This quick PR makes sure that a certain CSS margin rule for the "Build My Declaration" button on the EFNY homepage only applies to buttons inside the top level content container, and not anywhere else on the page where the rule isn't needed. This fixes an issue on mobile where the sticky footer button was improperly spaced because of this rule.