Update FullStory script tag (#2204)
This is a simple update of the script tag that FullStory provides to enable their tool on our site. We got this notification from FullStory back in the early summer: "We are preparing to update some of our infrastructure, and as a result if you haven’t updated your snippet since October 30, 2019 you will need to update the FullStory snippet on your site by July 13, 2021 to avoid a disruption in session recording."

For future reference, we were able to retrieve the new snippet by logging in to our organizational FullStory account and going to the "FullStory Setup" section of settings.