On the password verify page, ask user for their phone number if it isn't saved in the session (#2205)
This PR fixes a bug where users would arrive at the /en/declaration/password/verify page on EFNY or the /en/letter/password/verify page on NoRent without ever entering their phone number. In order to verify their password, this page assumes that the user has their phone number saved in the session to send them a verification code. However, in these odd circumstances where users arrive to this page directly without completing the earlier steps, this is not the case, and an error is thrown.