Add modal popup to EFNY to signal to users that the tool is deprecated (#2246)
This PR suspends the ability to send a hardship declaration on EFNY, and communicates to users that the tool is suspended and what other resources they have.

This PR largely makes use of the feature flag implemented in #2192 (the IS_EFNY_SUSPENDED environment variable). However, in that PR, we chose to update content on the homepage itself and then redirect all internal pages to the homepage. In this update, we keep all pages as is, and instead show a modal on all pages (except those needed to log in to an existing account).