Add new custom issue checklist to LA Habitability flow (#2265)
This PR adds the data structures and front-end scaffolding for the issue checklist on the LA Letter Builder site. Specifically, this PR introduces three new common data types that comprise the set of issues LA habitability letter users can select for their letter:

LaIssueChoice (I.e. HEALTH__PEELING_PAINT)
This type represents a concatenation of an issue and its larger category/area. It is fully analogous to the IssueChoice type from LOC, which has identical syntax: HOME__NO_COLD_WATER.

LaIssueAreaChoice (I.e. HEALTH)
This type represents just a category/area that's used to categorize the whole list of issues. This type has a one-to-one relationship with the collection of all possible prefixes of LaIssueChoice. It is fully analogous to the IssueAreaChoice type from LOC.

LaIssueRoomChoice (I.e. BEDROOM)
This type represents the room that the issue is found in. This type is a fixed list of rooms and has no relationship/overlap with either of the other new types. There is no analogous type found in the LOC app.

In addition to creating these new common data types, this PR implements the full front end, with the exception of the issue counter.