Add new content to NoRent specifying income requirement (#2323)
This PR adds a few content changes (including a new page) to the NoRent flow so that our tool is compliant with new income restrictions for the State of California. Specifically, this pr:

updates content on the "know your rights" page for users in California
adds a new "Please confirm your household income" page before the user selects which months to send letters for
updates content on the letter preview page
updates content on the letter itself
removes an irrelevant FAQ question
Note that the "Please confirm your household income" page sits at /rent-periods, which is the old path for the month selector page. This is to make sure that even if users link back to this url, they will not bypass the new income eligibility confirmation.