add django app for GCE screener responses (#2465)
Adds a new app to save user responses on the standalone Good Cause Eviction screener tool. We set up the model for responses, and a single API endpoint for POST requests from the screener. On first the first request with address it creates a new record and returns the ID that's autogenerated in the DB. On subsequent requests with that ID included it will update the fields with new data.

Since all data being sent it directly controlled by us on the GCE screener side (no direct user input strings), I don't think we need extensive validation with forms.py, and it wasn't working as it does in wow), so we use a simple pydantic model with basic tyes. But in future if we wanted we can use pydantics validation options to directly match the django model (eg. options for yes/no/unsure, etc.)

Since we are not using the frontend of tenants2, we need to log errors for 500 responses so they get tracked by rollbar.

I have the GCE_ORIGN set in justfix_environment.py with the temporary one we are using, but we'll need to update this once we settle on a final domain for the screener.

Also, unrelated to this project, I added port forwarding for the postgres DB in docker-compose so we can connect locally (eg. via postico) for easier debugging.

Companion PR for gce-screener: JustFixNYC/gce-screener#39