allow multiple valid cors origins for gce (#2470)
Originally we just had a single origin that was allowed for GCE requests. Now we follow the same pattern as we use in WOW, specifying a list of valid origins and a list of regexes of origins in the django settings file. Now all the prod/demo/prviews links should work