allow new gce record starting from results page (#2477)
Normally we expect new GCE records to be created from the home page when you search an address, starting with all the required address fields. But it's also technically possible to link directly to a results page with all the search params but with user id missing - in this case we'd be creating a brand new record at this step. Previously we weren't sending all the other fields, including the required address ones, at this step from from the front-end. But that is fixed with: JustFixNYC/gce-screener#123

Here we then just also need to initialize the new gce record with only the address fields, and then update the values so that we can still decide whether to use _initial or _final versions of the 3 fields where we have these different versions.