Make all params optional for GCE upload endpoint (#2478)
* make all params optional for post

* initialize empty gce record if no id, then update fields

In the post data there are generic versions of the _initial and _final fields, so we can't initialize a new record directly with this data since it won't recognize those field names. Now that all fields are optional we can just initialize it empty, and then update to fill the values with the _initial or _final versions, the as as if we were updating a pre-existing record.

---------

Co-authored-by: Maxwell Austensen <m.austensen@gmail.com>