Add GCE letter CC field/feature and fixes logic in email sending (#2490)
Adds a new boolean field to letter object for whether the user should be CC'd on the email to the landlord.
update the letter sending logic, and our util function, to include the CC argument.
Also fix some errors in the email sending logic - a typo in user vs landlord details, and dealing with the array of additional emails.