GCE Letter Sender REST backend (#2489)
Adds a new django app for the backend of the GCE Letter sender, gceletter. While the fontend is part of the existing gce-screeener project, we use a new app separate from the existing gce one. This app adds models for the letter (inheriting from existing model used in LATAC) along with linked user_details and landlord_details models, all with 1:1 relationships. We don't use any of the existing tenant platform user onboarding. The app is a fairly simple REST API, and uses pydantic for data validation. We don't use any of the typical GraphQL setup used on projects with their frontends on tenants2.

Not everything here is completely final, and in particular we need to add more tests, but this is the bulk of the work and everything up to the point of our first round of internal QA, where we'll need the API live on demo for testing.

companion PR for frontend: JustFixNYC/gce-screener#171